opt for a functional layer instead of a class layer
